Moved postprocessing format out of the "verarbeitungsschritte" configuration into the mimetypes configuration.
