--- src/main/java/de/bayern/gdi/model/ProcessingFormat.java
@@ -1,67 +1,0 @@
-/*
- * DownloadClient Geodateninfrastruktur Bayern
- *
- * (c) 2016 GSt. GDI-BY (gdi.bayern.de)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package de.bayern.gdi.model;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
-
-/** 'Format' of processing step configuration. */
-@XmlRootElement(name = "Format")
-@XmlAccessorType(XmlAccessType.FIELD)
-public class ProcessingFormat {
-
-    @XmlAttribute(name = "formattyp")
-    private String type;
-
-    @XmlElement(name = "MimeType")
-    private String mimeType;
-
-    public ProcessingFormat() {
-    }
-
-    /**
-     * @return the type
-     */
-    public String getType() {
-        return type;
-    }
-
-    /**
-     * @param type the type to set
-     */
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    /**
-     * @return the mimeType
-     */
-    public String getMimeType() {
-        return mimeType;
-    }
-
-    /**
-     * @param mimeType the mimeType to set
-     */
-    public void setMimeType(String mimeType) {
-        this.mimeType = mimeType;
-    }
-}
